+++ examples/SwichIcon/BasicUsage.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import SwitchIcon from 'src/SwitchIcon';
+
+function BasicUsage() {
+    return (
+        <div>
+            <SwitchIcon />
+            <SwitchIcon state="on" />
+        </div>
+    );
+}
+
+export default BasicUsage;