+++ src/__tests__/Checkbox.test.js
@@ -22,14 +22,14 @@ describe('Pure <Checkbox>', () => {
         const wrapper = shallow(
             <PureCheckbox>Foo children</PureCheckbox>
         );
-        expect(wrapper.contains(<input type="checkbox" />));
+        expect(wrapper.containsMatchingElement(<input type="checkbox" />)).toBeTruthy();
     });
 
     it('renders <input> in icon wrapper before rowComp parts', () => {
         const wrapper = shallow(
             <PureCheckbox>Foo children</PureCheckbox>
         );
-        expect(wrapper.childAt(0).hasClass('ic-checkbox__icon-wrapper'));
+        expect(wrapper.childAt(0).hasClass('gyp-checkbox__icon-wrapper')).toBeTruthy();
         expect(wrapper.childAt(0).find('input').exists()).toBeTruthy();
     });